% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_allVlns.R
\name{generate_allVlns}
\alias{generate_allVlns}
\title{Generate Violin Plots for given genes across different clusters}
\usage{
generate_allVlns(SeuFile, gene_list)
}
\arguments{
\item{SeuFile}{A Seurat object that contains the gene expression data.}

\item{gene_list}{A list of genes for which the violin plots need to be generated.}
}
\value{
Generates a PDF file containing the violin plots for the provided genes.
}
\description{
This function generates violin plots for each gene provided in the gene_list parameter across
different clusters in a Seurat object. It arranges these plots and saves them as a PDF file.
}
