% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_scaled_FPs.R
\name{generate_scaled_FPs}
\alias{generate_scaled_FPs}
\title{Generate Scaled Feature Plots}
\usage{
generate_scaled_FPs(
  seurat_objects,
  genes,
  filename = as.character(Sys.Date()),
  separate = F
)
}
\arguments{
\item{seurat_objects}{A named list of Seurat objects.}

\item{genes}{A character vector of gene names for which to generate feature plots.}

\item{filename}{The filename (without extension) to use for the output PDF files.}

\item{separate}{A boolean indicating whether to save a separate PDF for each Seurat object.
Default is TRUE.}
}
\value{
This function does not return a value. It generates PDF files with feature plots.
}
\description{
This function generates feature plots for a list of genes in multiple Seurat objects,
scaling the plots by the maximum value of each gene across all objects.
The generated plots are then saved as PDF files in an output directory.
}
