% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cells_from_integrated.R
\name{plot_cells_from_integrated}
\alias{plot_cells_from_integrated}
\title{Plot specific cells from an integrated Seurat object}
\usage{
plot_cells_from_integrated(integrated, ClusterName, sample)
}
\arguments{
\item{integrated}{A Seurat object. This is an integrated or merged Seurat object
that includes data from multiple samples.}

\item{ClusterName}{A character string specifying the name of the cell type or
cluster whose cells are to be visualized.}

\item{sample}{A Seurat object that contains scRNA-seq data for a specific
sample.}
}
\value{
This function does not return a value, but produces a UMAP plot for
        the selected cells from the integrated Seurat object.
}
\description{
This function generates a UMAP plot for specific cells in a given sample that
belong to a certain cluster from an integrated Seurat object.
}
\seealso{
\code{\link[Seurat]{DimPlot}}
}
