% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cells_from_sample.R
\name{plot_cells_from_sample}
\alias{plot_cells_from_sample}
\title{Plot cells of a specified sample and cluster type}
\usage{
plot_cells_from_sample(sample, ClusterName, integrated)
}
\arguments{
\item{sample}{A Seurat object that contains the sample data.}

\item{ClusterName}{A string that specifies the cell type or cluster name to be plotted.}

\item{integrated}{An integrated Seurat object that contains the merged data from multiple samples.}
}
\value{
The function does not return anything. It prints a UMAP plot to the console.
}
\description{
This function plots cells of a specified sample and cluster type using a UMAP plot.
}
\seealso{
\code{\link[Seurat]{DimPlot}}
}
