% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_qc_file.R
\name{save_qc_file}
\alias{save_qc_file}
\title{Save QC file}
\usage{
save_qc_file(
  SeuFile,
  version = "v1.0.0",
  qc.plot = TRUE,
  Heatmap = FALSE,
  csv = FALSE,
  pdf.height = 10,
  pdf.width = 12
)
}
\arguments{
\item{SeuFile}{Seurat object that contains the data.}

\item{version}{A string that provides the version number for the QC file. Default is "v1.0.0".}

\item{qc.plot}{Logical, if TRUE the function will generate QC plots for each cluster. Default is TRUE.}

\item{Heatmap}{Logical, if TRUE the function will generate a heatmap. Default is FALSE.}

\item{csv}{Logical, if TRUE the function will write CSV files for the top 10 markers and all markers per cluster. Default is FALSE.}

\item{pdf.height}{A numeric value specifying the height of the PDF. Default is 10.}

\item{pdf.width}{A numeric value specifying the width of the PDF. Default is 12.}
}
\value{
The function does not return anything. It saves the PDF and CSV files in the "output" directory.
}
\description{
This function saves UMAP plot and quality control (QC) files, with options to include QC plots and heatmaps, to output folder.
}
\seealso{
\code{\link[Seurat]{DimPlot}}, \code{\link[Seurat]{VlnPlot}}, \code{\link[Seurat]{DoHeatmap}}
}
